backend/clubmember/requirements.txt
asgiref==3.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
icalendar==6.3.1
idna==3.11
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2023.3
qrcode==8.2
requests==2.32.3
six==1.17.0
sqlparse==0.4.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.4.0
gunicorn==21.2.0
dj-database-url==2.1.0
sendgrid==6.11.0
google-generativeai==0.8.3
